wmvcore: Move IWMSyncReader implementation to separate file.
(cherry picked from commit 477dc82c4466b073af9a091ecd58dda83b3219ee)
